+++ massa-grpc/src/handler.rs
@@ -19,6 +19,7 @@ use crate::public::{
 
 #[cfg(feature = "execution-trace")]
 use crate::public::{get_operation_abi_call_stacks, get_slot_abi_call_stacks, get_slot_transfers};
+use crate::stream::new_blocks_server::new_blocks_server;
 use crate::stream::new_slot_execution_outputs_server::new_slot_execution_outputs_server;
 #[cfg(feature = "execution-trace")]
 use crate::stream::new_slot_transfers::new_slot_transfers;
@@ -243,6 +244,17 @@ impl grpc_api::public_service_server::PublicService for MassaPublicGrpc {
         request: tonic::Request<tonic::Streaming<grpc_api::NewBlocksRequest>>,
     ) -> Result<tonic::Response<Self::NewBlocksStream>, tonic::Status> {
         Ok(tonic::Response::new(new_blocks(self, request).await?))
+    }
+
+    type NewBlocksServerStream = NewBlocksStreamType;
+    /// handler for subscribe new blocks unidirectional
+    async fn new_blocks_server(
+        &self,
+        request: tonic::Request<grpc_api::NewBlocksRequest>,
+    ) -> Result<tonic::Response<Self::NewBlocksServerStream>, tonic::Status> {
+        Ok(tonic::Response::new(
+            new_blocks_server(self, request).await?,
+        ))
     }
 
     type NewEndorsementsStream = NewEndorsementsStreamType;
