+++ massa-grpc/src/stream/mod.rs
@@ -2,6 +2,8 @@
 
 /// stream new blocks
 pub mod new_blocks;
+/// unidirectional streaming new blocks
+pub mod new_blocks_server;
 /// stream new endorsements
 pub mod new_endorsements;
 /// stream new blocks headers
