+++ massa-grpc/src/stream/tools.rs
@@ -1,11 +1,19 @@
+use std::collections::HashSet;
+use std::str::FromStr;
+
+use crate::SlotRange;
 use crate::{config::GrpcConfig, error::GrpcError};
 use massa_execution_exports::{ExecutionOutput, SlotExecutionOutput};
+use massa_models::address::Address;
+use massa_models::block::SecureShareBlock;
+use massa_models::block_id::BlockId;
+use massa_models::slot::Slot;
 use massa_proto_rs::massa::api::v1::{self as grpc_api, NewSlotExecutionOutputsRequest};
 use massa_proto_rs::massa::model::v1::{self as grpc_model};
 
 /// Type declaration for NewSlotExecutionOutputsFilter
 #[derive(Clone, Debug, Default)]
-pub(crate) struct Filter {
+pub(crate) struct FilterNewSlotExec {
     // Execution output status to filter
     status_filter: Option<i32>,
     // Slot range to filter
@@ -23,10 +31,10 @@ pub(crate) struct Filter {
 }
 
 // This function returns a filter from the request
-pub(crate) fn get_filter(
+pub(crate) fn get_filter_slot_exec_out(
     request: NewSlotExecutionOutputsRequest,
     grpc_config: &GrpcConfig,
-) -> Result<Filter, GrpcError> {
+) -> Result<FilterNewSlotExec, GrpcError> {
     if request.filters.len() as u32 > grpc_config.max_filters_per_request {
         return Err(GrpcError::InvalidArgument(format!(
             "too many filters received. Only a maximum of {} filters are accepted per request",
@@ -34,7 +42,7 @@ pub(crate) fn get_filter(
         )));
     }
 
-    let mut result = Filter::default();
+    let mut result = FilterNewSlotExec::default();
 
     for query in request.filters.into_iter() {
         if let Some(filter) = query.filter {
@@ -74,7 +82,7 @@ pub(crate) fn get_filter(
 /// Return if the slot execution outputs should be send to client
 pub(crate) fn filter_map(
     slot_execution_output: SlotExecutionOutput,
-    filters: &Filter,
+    filters: &FilterNewSlotExec,
     grpc_config: &GrpcConfig,
 ) -> Option<SlotExecutionOutput> {
     match slot_execution_output {
@@ -98,7 +106,7 @@ pub(crate) fn filter_map(
 // Return if the execution outputs should be send and remove the fields that are not needed
 fn filter_map_exec_output_inner(
     mut exec_output: ExecutionOutput,
-    filters: &Filter,
+    filters: &FilterNewSlotExec,
     _grpc_config: &GrpcConfig,
     exec_status: i32,
 ) -> Option<ExecutionOutput> {
@@ -289,4 +297,135 @@ fn filter_map_exec_output_inner(
     }
 
     Some(exec_output)
+}
+
+// Type declaration for NewBlocksFilter
+#[derive(Clone, Debug)]
+pub(crate) struct FilterNewBlocks {
+    // Block ids to filter
+    block_ids: Option<HashSet<BlockId>>,
+    // Addresses to filter
+    addresses: Option<HashSet<Address>>,
+    // Slot range to filter
+    slot_ranges: Option<HashSet<SlotRange>>,
+}
+
+// This function returns a filter from the request
+pub(crate) fn get_filter_new_blocks(
+    request: grpc_api::NewBlocksRequest,
+    grpc_config: &GrpcConfig,
+) -> Result<FilterNewBlocks, GrpcError> {
+    if request.filters.len() as u32 > grpc_config.max_filters_per_request {
+        return Err(GrpcError::InvalidArgument(format!(
+            "too many filters received. Only a maximum of {} filters are accepted per request",
+            grpc_config.max_filters_per_request
+        )));
+    }
+
+    let mut block_ids_filter: Option<HashSet<BlockId>> = None;
+    let mut addresses_filter: Option<HashSet<Address>> = None;
+    let mut slot_ranges_filter: Option<HashSet<SlotRange>> = None;
+
+    // Get params filter from the request.
+    for query in request.filters.into_iter() {
+        if let Some(filter) = query.filter {
+            match filter {
+                grpc_api::new_blocks_filter::Filter::BlockIds(ids) => {
+                    if ids.block_ids.len() as u32 > grpc_config.max_block_ids_per_request {
+                        return Err(GrpcError::InvalidArgument(format!(
+                            "too many block ids received. Only a maximum of {} block ids are accepted per request",
+                            grpc_config.max_block_ids_per_request
+                        )));
+                    }
+
+                    let block_ids = block_ids_filter.get_or_insert_with(HashSet::new);
+                    for block_id in ids.block_ids {
+                        block_ids.insert(BlockId::from_str(&block_id).map_err(|_| {
+                            GrpcError::InvalidArgument(format!("invalid block id: {}", block_id))
+                        })?);
+                    }
+                }
+                grpc_api::new_blocks_filter::Filter::Addresses(addrs) => {
+                    if addrs.addresses.len() as u32 > grpc_config.max_addresses_per_request {
+                        return Err(GrpcError::InvalidArgument(format!(
+                            "too many addresses received. Only a maximum of {} addresses are accepted per request",
+                         grpc_config.max_addresses_per_request
+                        )));
+                    }
+
+                    let addresses = addresses_filter.get_or_insert_with(HashSet::new);
+                    for address in addrs.addresses {
+                        addresses.insert(Address::from_str(&address).map_err(|_| {
+                            GrpcError::InvalidArgument(format!("invalid address: {}", address))
+                        })?);
+                    }
+                }
+                grpc_api::new_blocks_filter::Filter::SlotRange(s_range) => {
+                    let slot_ranges = slot_ranges_filter.get_or_insert_with(HashSet::new);
+                    if slot_ranges.len() as u32 > grpc_config.max_slot_ranges_per_request {
+                        return Err(GrpcError::InvalidArgument(format!(
+                            "too many slot ranges received. Only a maximum of {} slot ranges are accepted per request",
+                         grpc_config.max_slot_ranges_per_request
+                        )));
+                    }
+
+                    let start_slot = s_range.start_slot.map(|s| s.into());
+                    let end_slot = s_range.end_slot.map(|s| s.into());
+
+                    let slot_range = SlotRange {
+                        start_slot,
+                        end_slot,
+                    };
+                    slot_range.check()?;
+                    slot_ranges.insert(slot_range);
+                }
+            }
+        }
+    }
+
+    Ok(FilterNewBlocks {
+        block_ids: block_ids_filter,
+        addresses: addresses_filter,
+        slot_ranges: slot_ranges_filter,
+    })
+}
+
+// This function checks if the block should be sent
+pub(crate) fn should_send_new_blocks(
+    signed_block: &SecureShareBlock,
+    filters: &FilterNewBlocks,
+    grpc_config: &GrpcConfig,
+) -> bool {
+    if let Some(block_ids) = &filters.block_ids {
+        if !block_ids.contains(&signed_block.id) {
+            return false;
+        }
+    }
+
+    if let Some(addresses) = &filters.addresses {
+        if !addresses.contains(&signed_block.content_creator_address) {
+            return false;
+        }
+    }
+
+    if let Some(slot_ranges) = &filters.slot_ranges {
+        let mut start_slot = Slot::new(0, 0); // inclusive
+        let mut end_slot = Slot::new(u64::MAX, grpc_config.thread_count - 1); // exclusive
+
+        for slot_range in slot_ranges {
+            start_slot = start_slot.max(slot_range.start_slot.unwrap_or_else(|| Slot::new(0, 0)));
+            end_slot = end_slot.min(
+                slot_range
+                    .end_slot
+                    .unwrap_or_else(|| Slot::new(u64::MAX, grpc_config.thread_count - 1)),
+            );
+        }
+        end_slot = end_slot.max(start_slot);
+        let current_slot = signed_block.content.header.content.slot;
+
+        return current_slot >= start_slot // inclusive
+            && current_slot < end_slot; // exclusive
+    }
+
+    true
 }
